More type definition safe
